Added admin to project roles
