+++ TaskManagementAPI/TaskManagementAPI/Services/Interfaces/IUserService.cs
@@ -4,8 +4,9 @@ namespace TaskManagementAPI.Services.Interfaces
 {
     public interface IUserService
     {
-        Task GetCurrentUserAsync(int id);
+        Task<User?> GetCurrentUserAsync(int id);
         Task EditPasswordAsync(string password, string userId);
+
 
     }
 }
